Mac stuff should only load on macs now
In each source loop, I check for osx directory and if the os is a mac.
It's expensive but it allows the convetion to remain simple. I thought
about giving each of them their own extension but that got more
complicated when considering maintain support for bash. Speaking of bash
is kinda broken right now.