path_unshift now moves path to front of $PATH
This was causing an issue cause /etc/path had /usr/local/bin but I
wanted it in front of /usr/bin. Since the if was breaking when it saw
/usr/local/bin was already there, it never inserted it before /usr/bin.
This works but now creates a race condition for path statements based
on the order the file gets sourced (file name). Still not crazy about
this.